GREAT DAY (AND FUN)
Today we programed a madlib, and a candy crush thingy.  I had a lot of fun with the code, and I put many weird words in the madlib. One hour and 30 minutes is a long time, but it passes so fast when you are having fun. HAVE A GOOD DAY.